+++ CMakeLists.txt
@@ -38,6 +38,7 @@ include(cmake/camera.cmake)
 include(cmake/encoder.cmake)
 include(cmake/decoder.cmake)
 include(cmake/display.cmake)
+include(cmake/bms.cmake)
 # include(cmake/location.cmake)
 
 
@@ -63,6 +64,12 @@ if (FEATURE_DISPLAY)
     include_directories(${glad_dir}/include)
     # for some reason, can only get this to build if I directly include glad instead of compiling it as a shared lib
     add_subdirectory(${internal_dir}/infrastructure/graphics)
+endif()
+if (FEATURE_BMS)
+    add_subdirectory(${internal_dir}/infrastructure/bms)
+endif()
+if (FEATURE_GPIO)
+    add_subdirectory(${internal_dir}/infrastructure/gpio)
 endif()
 if (FEATURE_LOCATION)
     include_directories(${third_party_dir}/libserialport)
