+++ internal/infrastructure/bms/bms.hpp
@@ -1,0 +1,8 @@
+//
+// Created by brucegoose on 5/26/23.
+//
+
+#ifndef INFRASTRUCTURE_BMS_HPP
+#define INFRASTRUCTURE_BMS_HPP
+
+#endif //INFRASTRUCTURE_BMS_HPP
