+++ internal/infrastructure/bms/serial_bms.cpp
@@ -1,0 +1,5 @@
+//
+// Created by brucegoose on 5/26/23.
+//
+
+#include "serial_bms.hpp"
