+++ internal/infrastructure/bms/serial_bms.hpp
@@ -1,0 +1,10 @@
+//
+// Created by brucegoose on 5/26/23.
+//
+
+#ifndef INFRASTRUCTURE_BMS_SERIAL_BMS_HPP
+#define INFRASTRUCTURE_BMS_SERIAL_BMS_HPP
+
+#include "bms.hpp"
+
+#endif //INFRASTRUCTURE_BMS_SERIAL_BMS_HPP
