+++ internal/infrastructure/gpio/pigpio_gpio.hpp
@@ -1,0 +1,10 @@
+//
+// Created by brucegoose on 5/26/23.
+//
+
+#ifndef INFRASTRUCTURE_GPIO_PIGPIO_GPIO_HPP
+#define INFRASTRUCTURE_GPIO_PIGPIO_GPIO_HPP
+
+#include "gpio.hpp"
+
+#endif //INFRASTRUCTURE_GPIO_PIGPIO_GPIO_HPP
