+++ internal/infrastructure/tcp/tcp_client.hpp
@@ -11,7 +11,7 @@
 #include <queue>
 
 #include "utils/buffers.hpp"
-#include "tcp_context.hpp"
+#include "internal/utils/asio_context.hpp"
 #include "tcp_utils.hpp"
 
 using boost::asio::ip::tcp;
