+++ internal/infrastructure/tcp/tcp_server.hpp
@@ -8,7 +8,7 @@
 #include <memory>
 #include <queue>
 
-#include "tcp_context.hpp"
+#include "internal/utils/asio_context.hpp"
 #include "utils/buffers.hpp"
 #include "tcp_utils.hpp"
