+++ internal/service/headset_streamer.cpp
@@ -17,9 +17,9 @@ namespace service {
     }
 
     void HeadsetStreamer::initialize(const service::HeadsetStreamerConfig &config) {
-        _tcp_context = infrastructure::TcpContext::Create(config);
+        _asio_context = AsioContext::Create(config);
         auto self(shared_from_this());
-        _tcp_client = infrastructure::TcpClient::Create(config, _tcp_context->GetContext(), std::move(self));
+        _tcp_client = infrastructure::TcpClient::Create(config, _asio_context->GetContext(), std::move(self));
         _graphics = infrastructure::Graphics::Create(config);
         self = shared_from_this();
         _decoder = infrastructure::Decoder::Create(
